+++ src/graphics.rs
@@ -7,7 +7,7 @@
 //!
 //! For loading and drawing images, to the screen, use [`Image`].
 //!
-//! [`run`]: crate::lifecycle::run
+//! [`run`]: crate::run::run
 
 use crate::QuicksilverError;
 
@@ -31,6 +31,7 @@ pub use self::surface::Surface;
 pub use self::vertex::{Element, Vertex};
 
 use crate::geom::*;
+use crate::Window;
 use golem::*;
 use std::iter;
 use std::mem::size_of;
@@ -527,7 +528,7 @@ impl Graphics {
     }
 
     /// Send the draw data to the GPU and paint it to the Window
-    pub fn present(&mut self, win: &blinds::Window) -> Result<(), QuicksilverError> {
+    pub fn present(&mut self, win: &Window) -> Result<(), QuicksilverError> {
         self.flush(None)?;
         win.present();
 
@@ -545,8 +546,8 @@ impl Graphics {
     /// The units given are physical units, not logical units. As such when using [`Window::size`],
     /// be sure to multiply by [`Window::scale_factor`].
     ///
-    /// [`Window::size`]: crate::lifecycle::Window::size
-    /// [`Window::scale_factor`]: crate::lifecycle::Window::scale_factor
+    /// [`Window::size`]: crate::Window::size
+    /// [`Window::scale_factor`]: crate::Window::scale_factor
     pub fn set_viewport(&self, x: u32, y: u32, width: u32, height: u32) {
         self.ctx.set_viewport(x, y, width, height);
     }
@@ -565,7 +566,7 @@ impl Graphics {
     }
 
     /// Set the viewport to cover the window, taking into account DPI
-    pub fn fit_to_window(&self, window: &blinds::Window) {
+    pub fn fit_to_window(&self, window: &Window) {
         let size = window.size();
         let scale = window.scale_factor();
         let width = size.x * scale;